github-testing/requirements.txt
ddt==1.6.0
jproperties==2.1.1
pip==22.3
py==1.11.0
requests==2.28.1
selenium==4.5.0
setuptools==60.2.0
wheel==0.37.1
PyAutoGUI~=0.9.53
pytest~=7.1.3
msedge-selenium-tools==3.141.4
webdriver-manager==3.8.4
allure-pytest==2.11.1
Pillow==9.3.0
pathlib==1.0.1